Introduction

This package serves to provide the necessary tools to perform net survival analysis, a branch of survival analysis dedicated to estimating the probability of survival from a particular event of interest compared to the general public. Some key features in NetSurvival.jl are:

Fitting different non-parametric estimators (Pohar Perme[1], Ederer II, ...)

Applying Grafféo's log-rank test[2] on different groups, including stratified covariables

...

Installation

The package is available on Julia's general registry, and can be installed either with the command Pkg.add("NetSurvival") or via the Pkg REPL mode:

] add NetSurvival

1 == 1

true

a = 2

2

a

2

NetSurvival.PoharPerme

NetSurvival.PoharPerme — Type

PoharPerme

BlaBlaBLa

[1]

M. P. Perme, J. Stare and J. Estève. On Estimation in Relative Survival. Biometrics 68, 113–120 (2011).

[2]

N. Grafféo, F. Castell, A. Belot and R. Giorgi. A Log-Rank-Type Test to Compare Net Survival Distributions. Biometrics 72, 760–769 (2016).