Getting Started

Fitting the non parametric estimators

NetSurvival.PoharPerme — Type

PoharPerme

This method estimates net survival probabilities by applying the following estimation:

\[\partial\hat{\Lambda}_E(t) = \frac{\sum_i \frac{dN_i(u)}{S_{P_i}(u)} - \sum_i \frac{Y_i(u)}{S_{P_i}(u)}d\Lambda_{P_i}(u)}{\sum_i \frac{Y_i(u)}{S_{P_i}(u)}}\]

To fit the Pohar Perme to your data based on a certain rate table, apply the example below to your code :

fit(PoharPerme, @formula(Surv(time,status)~covariable1 + covariable2), data, ratetable)

References:

[4] Perme, Maja Pohar and Stare, Janez and Estève, Jacques (2012). On Estimation in Relative Survival.

Example

We will illustrate with an example using the dataset colrec, which comprises $5971$ patients diagnosed with colon or rectal cancer between 1994 and 2000. This dataset is sourced from the Slovenia cancer registry. Given the high probability that the patients are Slovenian, we will be using the Slovenian mortality table slopop as reference for the populational rates. Subsequently, we can apply various non-parametric estimators for net survival analysis.

N.B.

Mortality tables may vary in structure, with options such as the addition or removal of specific covariates. To confirm that the mortality table is in the correct format, please refer to the documentation of RateTables.jl, or directly extract it from there.

By examining slopop, we notice it contains information regarding age and year, as expected for mortality tables. Additionally, it incorporates the covariate sex, which has two possible entries (:male or :female).

Pohar Perme

using NetSurvival, RateTables
pp1 = fit(PoharPerme, @formula(Surv(time,status)~1), colrec, slopop)

8149×5 DataFrame
  Row │ Sₑ          ∂Λₑ            σₑ             lower_95_CI  upper_95_CI 
      │ Float64     Float64        Float64        Float64      Float64     
──────┼────────────────────────────────────────────────────────────────────
    1 │   0.997105    0.00289493     0.000710632    0.995717      0.998495
    2 │   0.995214    0.0018967      0.0009186      0.993424      0.997007
    3 │   0.992149    0.00307987     0.00117581     0.989865      0.994438
    4 │   0.988748    0.00342788     0.0014077      0.986024      0.99148
    5 │   0.986016    0.00276316     0.0015717      0.982983      0.989058
    6 │   0.983283    0.00277109     0.00172089     0.979972      0.986605
    7 │   0.981722    0.00158763     0.0018035      0.978258      0.985199
    8 │   0.976973    0.00483737     0.00202472     0.973104      0.980858
  ⋮   │     ⋮             ⋮              ⋮             ⋮            ⋮
 8143 │   0.390001   -0.00054201     0.740969       0.0912751     1.6664
 8144 │   0.390212   -0.00054201     0.740969       0.0913245     1.6673
 8145 │   0.390424   -0.00054201     0.740969       0.091374      1.66821
 8146 │   0.390636   -0.00054201     0.740969       0.0914235     1.66911
 8147 │   0.390847   -0.00054201     0.740969       0.0914731     1.67002
 8148 │   0.391059   -0.00054201     0.740969       0.0915227     1.67092
 8149 │ NaN         NaN            NaN            NaN           NaN
                                                          8134 rows omitted

Applying the Grafféo log-rank test

NetSurvival.GraffeoTest — Type

GraffeoTest

The Grafféo test is a log-rank type test and is typically used in net survival analysis to determine the impact of certain covariates in the study.

The null $(H_0)$ hypothesis tests the following assumption:

\[\forall t \in [0,T], \Lambda_{E,g_1}(t) = \Lambda_{E,g_2}(t) = ... = \Lambda_{E,g_k}(t)\]

where $G = {g_1,...,g_k}$ is a partition of $1,...,n$ consisting of disjoint groups of individuals that we wish to compare to each other. For all group $g \in G$, let's denote the numerator and denominator of the Pohar Perme (partial) excess hazard estimators, restricted to individuals in the group, by:

$\partial N_{E,g}(s) = \sum_{i \in g} \frac{\partial N_i(s)}{S_{P_i}(s)} - \frac{Y_i(s)}{S_{P_i}(s)}\partial\Lambda_{P_i}(s)$

$Y_{E,g}(s) = \sum_{i \in g} \frac{Y_i(s)}{S_{P_i}(s)}$

$R_{g}(s) = \frac{Y_{E,g}(s)}{\sum_{g\in G} Y_{E,g}(s)}$

Then, define the vector $\mathbf Z = \left(Z_{g_r}: r \in 1,...,k-1 \right)$ with entries:

\[Z_g(T) = N_{E,g}(s) - \int_{0}^T Y_{E,g}(s) \partial\hat{\Lambda}_E(s)\]

The test statistic is then given by:

\[U(T) = \mathbf Z(T)'\hat{\Sigma}_Z^{-1} \mathbf Z(T)\]

where the entries of the $\hat{\Sigma}_Z$ matrix are given by:

\[\sigma_{g,h}(T) = \int_0^T \sum_{\ell \in G} \left(\delta_{g,\ell} - R_g(t) \right)\left(\delta_{h,\ell} - R_h(t)\right) \left(\sum_{i\in\ell} \frac{\partial N_i(s)}{S^2_{P_i}}\right)\]

Under $H_0$, the statistic $U(T)$ is asymptotically $\chi^2(k-1)$-distributed.

To apply the test to your data based on a certain rate table, apply the example below to your code :

fit(GraffeoTest, @formula(Surv(time,status)~covariable1 + covariable2), data, ratetable)

If you wish to stratify a covariate:

fit(GraffeoTest, @formula(Surv(time,status)~covariable1 + Strata(covariable2)), data, ratetable)

References:

[5] Grafféo, Nathalie and Castell, Fabienne and Belot, Aurélien and Giorgi, Roch (2016). A Log-Rank-Type Test to Compare Net Survival Distributions.

Example

When applying the test to the same data as before, we get:

test1 = fit(GraffeoTest, @formula(Surv(time,status)~stage), colrec, slopop)

Grafféo's log-rank-type-test
1×3 DataFrame
 Row │ test_statistic  degrees_of_freedom  p_value      
     │ Float64         Int64               Float64      
─────┼──────────────────────────────────────────────────
   1 │        658.779                   3  1.81926e-142

The p-value is well under $0.05$, meaning that the different groups identified by the stage variable have different survival probabilities. Thus, it should be taken into consideration in the study.

test2 = fit(GraffeoTest, @formula(Surv(time,status)~sex), colrec, slopop)

Grafféo's log-rank-type-test
1×3 DataFrame
 Row │ test_statistic  degrees_of_freedom  p_value  
     │ Float64         Int64               Float64  
─────┼──────────────────────────────────────────────
   1 │       0.843748                   1  0.358327

For the sex variable, we notice that the p-value is above $0.05$ indicating that there isn't a difference between male and female patients.

test4 = fit(GraffeoTest, @formula(Surv(time,status)~stage+Strata(sex)), colrec, slopop)

Grafféo's log-rank-type-test
1×3 DataFrame
 Row │ test_statistic  degrees_of_freedom  p_value      
     │ Float64         Int64               Float64      
─────┼──────────────────────────────────────────────────
   1 │        678.514                   3  9.56849e-147